:truck: Move bar list to core
