+++ src/frontend/packages/core/src/components/BarList/BarList.module.css
